% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EigenRF.R
\name{rf_norm}
\alias{rf_norm}
\title{Title Metabolites Normalization Method:Eigen_RF}
\usage{
rf_norm(peak, groups, metabolites)
}
\arguments{
\item{peak}{the table of peak values}

\item{groups}{the vector of groups}

\item{metabolites}{the table or vector of metabolites}
}
\value{
the normalized peak table
}
\description{
Title Metabolites Normalization Method:Eigen_RF
}
\examples{
Eigen_RF(peak, groups, metabolites)
\donttest {
Eigen_RF(peak, groups, metabolites)
}
}
